chore: Extract logic to compute codegen TS extendsProps and props in getProps function
